+++ tools/test262regex/js_regex.cpp
@@ -282,7 +282,7 @@ static unsigned int toUTF16Offset(const StringRef ref, const unsigned int codePo
   const char *end = ref.begin() + codePointOffset;
   for (const char *iter = ref.begin(); iter != end;) {
     int codePoint;
-    if (unsigned int len = UnicodeUtil::utf8ToCodePoint(iter, end, codePoint); len) {
+    if (unsigned int len = UnicodeUtil::wtf8ToCodePoint(iter, end, codePoint); len) {
       iter += len;
     } else { // put dummy
       iter++;